支持浮点坐标并增强 MaskWindow 功能
- 将游戏内坐标从 (int, int) 改为 (double, double)，支持更高精度。
- 移除 OverlayService，新增 MaskWindow，支持鼠标坐标显示、动态更新位置及图形显示开关。
- 调整 GameContext 坐标转换逻辑，支持浮点数输入。
- 脚本引擎支持浮点坐标，新增脚本模型版本转换逻辑 (1.0 -> 1.1)。
- 更新 UI：移除 OverlayForm，改用 MaskWindow 显示鼠标坐标。
- 策略逻辑默认启用 MaskWindow，并根据设置控制图形显示。
- 优化 Instruction 坐标显示逻辑，新增智能文本框定位功能。

+++ BTD6AutoCommunity/GameObjects/HeroClass.cs
@@ -11,7 +11,7 @@ namespace BTD6AutoCommunity.GameObjects
     public class HeroClass
     {
         public string name;  // 猴子塔的名字
-        public (int, int) coordinates  = (0, 0);  // 部署坐标 (x, y)
+        public (double, double) coordinates  = (0, 0);  // 部署坐标 (x, y)
 
         public bool exsitence;
         public bool storeOpen;
@@ -34,7 +34,7 @@ namespace BTD6AutoCommunity.GameObjects
             return hero;
         }
 
-        public bool PlaceHero((int, int) pos)
+        public bool PlaceHero((double, double) pos)
         {
             if (exsitence) return false;
             coordinates = pos;
@@ -49,7 +49,7 @@ namespace BTD6AutoCommunity.GameObjects
             return true;
         }
 
-        public (int X, int Y) GetCoordinates()
+        public (double X, double Y) GetCoordinates()
         {
             return coordinates;
         }
